OK-671: Refaktoroitu HakutoiveetTable erilliseen tiedostoon
